% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas_loss_reserve_db}
\alias{cas_loss_reserve_db}
\title{CAS Loss Reserve Triangle Database}
\format{a \code{\link[tibble:tibble]{tibble}} with columns as described below

Column descriptions:
\itemize{
\item line: name of line of business ("ppauto", "wkcomp", "comauto", "medmal", "prodliab", "othliab")
\item group_id: NAIC company code
\item company
\item acc_yr
\item dev_yr
\item dev_lag: development year (acc_yr - 1987 + dev_yr - 1987 - 1)
\item cum_incurred_loss
\item increm_incurred_loss
\item cum_paid_loss
\item increm_paid_loss
\item bulk_ibnr: bulk and IBNR reserves on net losses and defense and cost containment expenses reported at year end
\item direct_ep: direct and assumed earned premium
\item ceded_ep: earned premium ceded to reinsurers
\item net_ep: premium earned net of reinsurance
\item single_entity: 1 indicates a single entity, 0 indicates a group insurer
\item posted_reserve_1997: Posted reserves in year 1997 taken from the Underwriting and Investment Exhibit - Part 2A,
including net losses unpaid and unpaid loss adjustment expenses
}}
\source{
\href{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}{LOSS RESERVING DATA PULLED FROM NAIC SCHEDULE P}
}
\usage{
cas_loss_reserve_db
}
\description{
Full dataset from the Casualty Actuarial Society's (CAS) loss reserving
\href{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}{database}.

From the \href{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}{CAS website}:
"Our goal is to prepare a clean and nice data set of loss triangles that could be used for claims reserving studies.
The data includes major personal and commercial lines of business from U.S. property casualty insurers.
The claims data comes from Schedule P - Analysis of Losses and Loss Expenses
in the National Association of Insurance Commissioners (NAIC) database."

This package pulls the by line .csv files from the CAS website and creates a \href{http://vita.had.co.nz/papers/tidy-data.html}{tidy}
data set. Column names were changed from the original data to make them more user friendly.
}
\keyword{datasets}
