% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\docType{data}
\name{custom_colors}
\alias{custom_colors}
\title{Default Color Scheme}
\format{An object of class \code{list} of length 19.}
\usage{
custom_colors
}
\description{
List of default custom colors
}
\details{
Convenience list of colors in #RGB format.  Just type
names(custom_colors) to see the list of colors included.
}
\examples{

\dontrun{ggplot() + geom_line(aes(x=-5:5, y=(-5:5)^2), colour=custom_colors$gold)}

}
\keyword{datasets}
