% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.MomentsToParams}
\alias{.MomentsToParams}
\title{Transform mean and cv into required distribution parameters}
\usage{
.MomentsToParams(param_list, dist = "lnorm")
}
\arguments{
\item{dist}{character with distribution abbreviation}

\item{params_list}{list with mean and cv}
}
\description{
Given a list of mean and cv and selected distribution,
transform mean and cv into necessary parameters to work with R dists.
}
\keyword{internal}
