% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-triangle-subsets.R
\name{get_meyers_subset}
\alias{get_meyers_subset}
\title{Subset CAS loss reserve database}
\usage{
get_meyers_subset(cas_triangle_database)
}
\arguments{
\item{cas_triangle_database}{the CAS loss reserve database \link{cas_loss_reserve_db}.}
}
\value{
a \code{tibble} filtered to include only 200 triangles used by Glenn Meyers
}
\description{
Subset CAS loss reserve database
}
\examples{
get_meyers_subset(cas_loss_reserve_db)

}
