% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meyers_2016_wintereforum_appendix}
\alias{meyers_2016_wintereforum_appendix}
\title{Appendix from Glenn Meyer's CAS E-Forum, Winter 2016 paper}
\format{a \code{\link[tibble:tibble]{tibble}}. See paper for description of the columns.}
\source{
\href{http://www.casact.org/pubs/forum/16wforum/Meyers.pdf}{Dependencies in Stochastic Loss Reserve Models}
\href{http://www.casact.org/pubs/forum/16wforum/02b_Meyers_Dependencies_Appendix-10-13-2015.xls}{Appendix spreadsheet}
}
\usage{
meyers_2016_wintereforum_appendix
}
\description{
The "Univariate Output" tab from the appendix of Glenn Meyer's paper
\href{http://www.casact.org/pubs/forum/16wforum/Meyers.pdf}{Dependencies in Stochastic Loss Reserve Models},
as pulled from the \href{http://www.casact.org/pubs/forum/16wforum/02b_Meyers_Dependencies_Appendix-10-13-2015.xls}{CAS website}.
}
\keyword{datasets}
