% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meyers_2019_appendix}
\alias{meyers_2019_appendix}
\title{Appendix from the Second Edition of Glenn Meyer's Bayesian Stochastic Loss Reserve CAS Monograph}
\format{a \code{\link[tibble:tibble]{tibble}}. See paper for description of the columns.}
\usage{
meyers_2019_appendix
}
\description{
The "Univariate Output" tab from the appendix of Glenn Meyer's paper
\href{https://www.casact.org/pubs/monographs/papers/08-Meyers.pdf}{Stochastic Loss Reserving Using Bayesian MCMC Models (2nd edition)},
as pulled from the \href{https://www.casact.org/pubs/monographs/index.cfm?fa=meyers-monograph08}{CAS website}.
}
\keyword{datasets}
