% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pct_format}
\alias{pct_format}
\title{Format number in percent style}
\usage{
pct_format(x, digits = 2, ...)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits to keep}

\item{...}{additional parameters passed to \link{formatC}}
}
\value{
character vector of formatted percentages
}
\description{
Format number in percent style
}
\examples{
pct_format(0.8546, digits = 2)

}
