% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservetestr-package.R
\docType{package}
\name{reservetestr-package}
\alias{reservetestr}
\alias{reservetestr-package}
\title{reservetestr: Provides a Framework for Testing Loss Reserve Methods}
\description{
This pacakge provides a framework for testing loss reserve methods. Specifically, an interface to test methods against the Casualty Actuarial Society (CAS) Loss Reserve Database is provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/problemofpoints/reservetestr}
  \item Report bugs at \url{https://github.com/problemofpoints/reservetestr/issues}
}

}
\author{
\strong{Maintainer}: Adam Troyer \email{problemofpoints@gmail.com}

}
\keyword{internal}
