% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestr.R
\name{run_single_backtest}
\alias{run_single_backtest}
\title{Run reserve backtest for a single reserving method}
\usage{
run_single_backtest(
  loss_data,
  reserving_function,
  lines_to_include = c("comauto", "othliab", "ppauto", "wkcomp"),
  loss_type_to_backtest = c("paid", "case", "ultimate"),
  by_year = FALSE,
  to_ultimate = TRUE,
  method_label = NULL,
  ...
)
}
\arguments{
\item{loss_data}{a nested \code{tibble} containing loss triangle data. See \link{cas_loss_reserve_db}.}

\item{reserving_function}{a function that takes as input a loss triangle and outputs estimated reserve}

\item{lines_to_include}{a character vector with the lines of business to include}

\item{loss_type_to_backtest}{a character indicating whether to use paid, case-incurred, or ultimate losses}

\item{by_year}{logical, whether to estimate results by year (TRUE) or in total across all years (FALSE). Default is FALSE.}

\item{to_ultimate}{logical, whether to estimate results to ultimate (TRUE) or for the next calendar year (FALSE). Default is TRUE.}

\item{method_label}{a character giving the name given to method. Default is "results"}

\item{...}{arguments passed on to the \code{reserving_function}}
}
\value{
a \code{tibble} with the results of the method and back-testing results as columns
}
\description{
Run reserve backtest for a single reserving method
}
\examples{
data <- cas_loss_reserve_db \%>\%
 get_meyers_subset()
data <- data[1:3,]

backtest_mack_paid <- run_single_backtest(data, testr_MackChainLadder, lines_to_include = "comauto",
method_label = "mack_paid")
}
