% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr-functions.R
\name{testr_MackChainLadder}
\alias{testr_MackChainLadder}
\title{Back-test MackChainLadder method}
\usage{
testr_MackChainLadder(
  train_data,
  test_data,
  loss_type = c("paid", "case", "ultimate"),
  .progress = NULL,
  ...
)
}
\arguments{
\item{train_data}{a \code{triangle} object containing upper left of a loss triangle. Used as input to \code{\link[ChainLadder:MackChainLadder]{ChainLadder::MackChainLadder()}}.}

\item{test_data}{a \code{triangle} object containing the lower right portion of loss triangle. Used to test model estimates.}

\item{loss_type}{a character indicating whether to use paid, case-incurred, or ultimate loss triangle}

\item{.progress}{a \code{Progress} object used to display a text based progress bar. Default is NULL.}

\item{...}{additional arguments passed on to \code{\link[ChainLadder:MackChainLadder]{ChainLadder::MackChainLadder()}}}
}
\value{
a \code{tibble} containing the true reserve, estimated mean, estimated coefficient of variation, and the implied
percentile of the actual outcome
}
\description{
Back-test MackChainLadder method
}
\examples{
train_data <- cas_loss_reserve_db \%>\%
 get_meyers_subset() \%>\%
 purrr::pluck("train_tri_set", 8)
test_data <- cas_loss_reserve_db \%>\%
 get_meyers_subset() \%>\%
 purrr::pluck("test_tri_set", 8)

testr_mack <- testr_MackChainLadder(train_data, test_data, loss_type = "paid")
}
